+++ packages/hydrogen/src/framework/Hydration/Html.tsx
@@ -16,11 +16,6 @@ export function Html({children, template, htmlAttrs, bodyAttrs}: HtmlOptions) {
       <head dangerouslySetInnerHTML={{__html: head}} />
       <body {...bodyAttrs}>
         <div id="root">{children}</div>
-        {/* In production, Vite bundles the entrypoint JS inside <head> */}
-        {/* @ts-ignore because module=commonjs doesn't allow this */}
-        {import.meta.env.DEV && (
-          <script type="module" src="/src/entry-client.jsx"></script>
-        )}
       </body>
     </html>
   );